+++ build-logic/build.gradle.kts
@@ -13,5 +13,4 @@ repositories {
 dependencies {
     implementation(libs.kotlin.gradle.plugin)
     implementation(libs.gradle)
-    implementation(libs.vanniktech.publish)
 }
