+++ build.gradle.kts
@@ -3,6 +3,7 @@
 plugins {
     alias(libs.plugins.compose.compiler) apply false
     alias(libs.plugins.jetbrains.compose) apply false
+    alias(libs.plugins.vanniktech.maven.publish) apply false
 }
 
 buildscript {
