+++ jujubasvg/version.properties
@@ -1,1 +1,1 @@
-VERSION=1.1.1
+VERSION=1.2.0
